adjusted code after rebase
