+++ libsql-server/src/namespace/meta_store.rs
@@ -13,8 +13,6 @@ use libsql_sys::wal::{
 };
 use parking_lot::Mutex;
 use prost::Message;
-use rusqlite::{params, OptionalExtension};
-use serde::{Deserialize, Serialize};
 use tokio::sync::oneshot;
 use tokio::sync::{
     mpsc,
@@ -23,8 +21,7 @@ use tokio::sync::{
 
 use crate::config::BottomlessConfig;
 use crate::connection::config::DatabaseConfig;
-use crate::connection::program::Program;
-use crate::schema::{MigrationJob, MigrationJobStatus, MigrationTask, MigrationTaskStatus};
+use crate::schema::{MigrationDetails, MigrationSummary};
 use crate::{
     config::MetaStoreConfig, connection::libsql::open_conn_active_checkpoint, error::Error, Result,
 };
@@ -236,267 +233,6 @@ impl MetaStoreInner {
         txn.commit()?;
 
         Ok(())
-    }
-
-    /// Create a migration task, and returns the jobn id
-    fn register_schema_migration_task(
-        &mut self,
-        schema: &NamespaceName,
-        migration: &Program,
-    ) -> Result<i64> {
-        let txn = self.conn.transaction()?;
-
-        // get the config for the schema and validate that it's actually a schema
-        let mut stmt =
-            txn.prepare("SELECT namespace, config FROM namespace_configs where namespace = ?")?;
-        let mut rows = stmt.query([schema.as_str()])?;
-        let Some(row) = rows.next()? else {
-            todo!("no such schema")
-        };
-        let config_bytes = row.get_ref(1)?.as_blob().unwrap();
-        let config = DatabaseConfig::from(&metadata::DatabaseConfig::decode(config_bytes)?);
-        if !config.is_shared_schema {
-            todo!("not a shared schema table");
-        }
-
-        drop(rows);
-
-        stmt.finalize()?;
-
-        let migration_serialized = serde_json::to_string(&migration).unwrap();
-        txn.execute(
-            "INSERT INTO migration_jobs (schema, migration, status) VALUES (?, ?, ?)",
-            (
-                schema.as_str(),
-                &migration_serialized,
-                MigrationJobStatus::WaitingDryRun as u64,
-            ),
-        )?;
-        let job_id = txn.last_insert_rowid();
-
-        txn.execute(
-            "
-            INSERT INTO
-                migration_job_pending_tasks (job_id, target_namespace, status)
-            SELECT job_id, namespace, status
-                FROM shared_schema_links 
-                CROSS JOIN (SELECT ? as job_id, ? as status)
-            WHERE shared_schema_name = ?",
-            (
-                job_id,
-                MigrationTaskStatus::Enqueued as u64,
-                schema.as_ref(),
-            ),
-        )?;
-
-        txn.commit()?;
-
-        Ok(job_id)
-    }
-
-    /// returns a batch of tasks for job_id that are in the passed status
-    fn get_next_pending_migration_tasks_batch(
-        &mut self,
-        job_id: i64,
-        status: MigrationTaskStatus,
-        limit: usize,
-    ) -> crate::Result<Vec<MigrationTask>> {
-        let txn = self
-            .conn
-            .transaction_with_behavior(rusqlite::TransactionBehavior::Immediate)?;
-        let tasks = txn
-            .prepare(
-                "SELECT task_id, target_namespace, status, job_id 
-                FROM migration_job_pending_tasks 
-                WHERE job_id = ? AND status = ? AND task_id NOT IN (select * from enqueued_tasks)
-                LIMIT ?",
-            )?
-            .query_map((job_id, status as u64, limit), |row| {
-                let task_id = row.get::<_, i64>(0)?;
-                let namespace = NamespaceName::from_string(row.get::<_, String>(1)?).unwrap();
-                let status = MigrationTaskStatus::from_int(row.get::<_, u64>(2)?);
-                let job_id = row.get::<_, i64>(3)?;
-                Ok(MigrationTask {
-                    namespace,
-                    status,
-                    job_id,
-                    task_id,
-                })
-            })?
-            .map(|r| r.map_err(Into::into))
-            .collect::<crate::Result<Vec<_>>>()?;
-
-        for task in tasks.iter() {
-            txn.execute("INSERT INTO enqueued_tasks VALUES (?)", [task.task_id])?;
-        }
-
-        txn.commit()?;
-        Ok(tasks)
-    }
-
-    fn update_task_status(
-        &mut self,
-        task: MigrationTask,
-        error: Option<String>,
-    ) -> crate::Result<()> {
-        assert!(error.is_none() || task.status.is_failure());
-        let txn = self.conn.transaction()?;
-        txn.execute(
-            "UPDATE migration_job_pending_tasks SET status = ?, error = ? WHERE task_id = ?",
-            (task.status as u64, error, task.task_id),
-        )?;
-        txn.commit()?;
-        Ok(())
-    }
-
-    /// Attempt to set the job to DryRunSuccess.
-    /// Checks that:
-    /// - current state is WaitinForDryRun
-    /// - all tasks are DryRunSuccess
-    fn job_step_dry_run_success(&self, mut job: MigrationJob) -> crate::Result<MigrationJob> {
-        let row_changed = self.conn.execute("
-            WITH tasks AS (SELECT * FROM migration_job_pending_tasks WHERE job_id = ?1)
-                UPDATE migration_jobs
-                SET status = ?2
-                WHERE job_id = ?1
-                    AND status = ?3
-                    AND (SELECT count(*) from tasks) = (SELECT count(*) FROM tasks WHERE status = ?4)",
-            (
-                job.job_id(),
-                MigrationJobStatus::DryRunSuccess as u64,
-                MigrationJobStatus::WaitingDryRun as u64,
-                MigrationTaskStatus::DryRunSuccess as u64))?;
-
-        if row_changed == 0 {
-            return Ok(job);
-        }
-
-        job.status = MigrationJobStatus::DryRunSuccess;
-
-        Ok(job)
-    }
-
-    fn update_job_status(&self, job_id: i64, status: MigrationJobStatus) -> crate::Result<()> {
-        self.conn.execute(
-            "UPDATE migration_jobs SET status = ? WHERE job_id = ?",
-            (status as u64, job_id),
-        )?;
-        Ok(())
-    }
-
-    fn get_migration_details(
-        &mut self,
-        schema: NamespaceName,
-        job_id: u64,
-    ) -> crate::Result<MigrationDetails> {
-        let status = self.conn.query_row(
-            "SELECT status
-            FROM migration_jobs
-            WHERE schema = ? AND job_id = ?",
-            params![schema.as_str(), job_id],
-            |r| {
-                let status: Option<u64> = r.get(0)?;
-                Ok(status.map(MigrationJobStatus::from_int))
-            },
-        )?;
-        let mut stmt = self.conn.prepare(
-            "SELECT target_namespace, status, error
-            FROM migration_job_pending_tasks
-            WHERE job_id = ?",
-        )?;
-        let rows = stmt.query([job_id])?.mapped(|r| {
-            let target_namespace = r.get(0)?;
-            let status: Option<u64> = r.get(1)?;
-            let error: Option<String> = r.get(2)?;
-            Ok(MigrationJobProgress {
-                namespace: target_namespace,
-                status: status.map(MigrationJobStatus::from_int),
-                error,
-            })
-        });
-        let mut progress = Vec::new();
-        for row in rows {
-            progress.push(row?);
-        }
-        Ok(MigrationDetails {
-            job_id,
-            status,
-            progress,
-        })
-    }
-
-    fn get_migrations_summary(&mut self, schema: NamespaceName) -> crate::Result<MigrationSummary> {
-        let schema_version: i64 = self
-            .conn
-            .query_row("PRAGMA schema_version;", (), |r| r.get(0))?;
-        let mut stmt = self.conn.prepare(
-            "SELECT job_id, status
-            FROM migration_jobs
-            WHERE schema = ?
-            ORDER BY job_id DESC",
-        )?;
-        let rows = stmt.query([schema.as_str()])?.mapped(|r| {
-            let status: Option<u64> = r.get(1)?;
-            Ok(MigrationJobSummary {
-                job_id: r.get(0)?,
-                status: status.map(MigrationJobStatus::from_int),
-            })
-        });
-        let mut migrations = Vec::new();
-        for row in rows {
-            migrations.push(row?);
-        }
-        Ok(MigrationSummary {
-            schema_version,
-            migrations,
-        })
-    }
-
-    fn get_next_pending_migration_job(&mut self) -> crate::Result<Option<MigrationJob>> {
-        let mut job = self
-            .conn
-            .query_row(
-                "SELECT job_id, status, migration, schema
-            FROM migration_jobs
-            WHERE status != ? AND status != ?
-            LIMIT 1",
-                (
-                    MigrationJobStatus::RunSuccess as u64,
-                    MigrationJobStatus::RunFailure as u64,
-                ),
-                |row| {
-                    let job_id = row.get::<_, i64>(0)?;
-                    let status = MigrationJobStatus::from_int(row.get::<_, u64>(1)?);
-                    let migration = serde_json::from_str(row.get_ref(2)?.as_str()?).unwrap();
-                    let schema = NamespaceName::from_string(row.get::<_, String>(3)?).unwrap();
-                    Ok(MigrationJob {
-                        schema,
-                        job_id,
-                        status,
-                        migration,
-                        progress: Default::default(),
-                    })
-                },
-            )
-            .optional()?;
-
-        if let Some(ref mut job) = job {
-            self.conn
-                .prepare(
-                    "
-                SELECT status, count(*)
-                FROM migration_job_pending_tasks 
-                WHERE job_id = ?
-                GROUP BY status",
-                )?
-                .query_map([job.job_id], |row| {
-                    job.progress[row.get::<_, usize>(0)?] = row.get::<_, usize>(1)?;
-                    Ok(())
-                })?
-                .collect::<Result<(), rusqlite::Error>>()?;
-        }
-
-        Ok(job)
     }
 
     #[tracing::instrument(skip(self))]
@@ -722,10 +458,12 @@ impl MetaStore {
         schema: NamespaceName,
     ) -> crate::Result<MigrationSummary> {
         let inner = self.inner.clone();
-        let summary =
-            tokio::task::spawn_blocking(move || inner.lock().get_migrations_summary(schema))
-                .await
-                .unwrap()?;
+        let summary = tokio::task::spawn_blocking(move || {
+            let mut lock = inner.lock();
+            crate::schema::get_migrations_summary(&mut lock.conn, schema)
+        })
+        .await
+        .unwrap()?;
         Ok(summary)
     }
 
@@ -735,88 +473,13 @@ impl MetaStore {
         job_id: u64,
     ) -> crate::Result<MigrationDetails> {
         let inner = self.inner.clone();
-        let summary =
-            tokio::task::spawn_blocking(move || inner.lock().get_migration_details(schema, job_id))
-                .await
-                .unwrap()?;
-        Ok(summary)
-    }
-
-    pub async fn register_schema_migration(
-        &self,
-        schema: NamespaceName,
-        migration: Arc<Program>,
-    ) -> crate::Result<i64> {
-        let inner = self.inner.clone();
-        let job_id = tokio::task::spawn_blocking(move || {
-            inner
-                .lock()
-                .register_schema_migration_task(&schema, &migration)
+        let summary = tokio::task::spawn_blocking(move || {
+            let mut lock = inner.lock();
+            crate::schema::get_migration_details(&mut lock.conn, schema, job_id)
         })
         .await
         .unwrap()?;
-
-        Ok(job_id)
-    }
-
-    pub(crate) async fn job_step_dry_run_success(
-        &self,
-        job: MigrationJob,
-    ) -> crate::Result<MigrationJob> {
-        let inner = self.inner.clone();
-
-        let job = tokio::task::spawn_blocking(move || inner.lock().job_step_dry_run_success(job))
-            .await
-            .unwrap()?;
-
-        Ok(job)
-    }
-
-    pub(crate) async fn update_task_status(
-        &self,
-        task: MigrationTask,
-        error: Option<String>,
-    ) -> crate::Result<()> {
-        let inner = self.inner.clone();
-        tokio::task::spawn_blocking(move || inner.lock().update_task_status(task, error))
-            .await
-            .unwrap()?;
-
-        Ok(())
-    }
-
-    pub(crate) async fn get_next_pending_job(&self) -> crate::Result<Option<MigrationJob>> {
-        let inner = self.inner.clone();
-        tokio::task::spawn_blocking(move || inner.lock().get_next_pending_migration_job())
-            .await
-            .unwrap()
-    }
-
-    pub(crate) async fn get_next_pending_migration_tasks_batch(
-        &self,
-        job_id: i64,
-        status: MigrationTaskStatus,
-        limit: usize,
-    ) -> crate::Result<Vec<MigrationTask>> {
-        let inner = self.inner.clone();
-        tokio::task::spawn_blocking(move || {
-            inner
-                .lock()
-                .get_next_pending_migration_tasks_batch(job_id, status, limit)
-        })
-        .await
-        .unwrap()
-    }
-
-    pub(crate) async fn update_job_status(
-        &self,
-        job_id: i64,
-        status: MigrationJobStatus,
-    ) -> crate::Result<()> {
-        let inner = self.inner.clone();
-        tokio::task::spawn_blocking(move || inner.lock().update_job_status(job_id, status))
-            .await
-            .unwrap()
+        Ok(summary)
     }
 }
 
@@ -909,188 +572,5 @@ impl MetaStoreHandle {
         };
 
         Ok(())
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct MigrationSummary {
-    pub schema_version: i64,
-    pub migrations: Vec<MigrationJobSummary>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct MigrationJobSummary {
-    pub job_id: u64,
-    pub status: Option<MigrationJobStatus>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct MigrationDetails {
-    pub job_id: u64,
-    pub status: Option<MigrationJobStatus>,
-    pub progress: Vec<MigrationJobProgress>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct MigrationJobProgress {
-    pub namespace: String,
-    pub status: Option<MigrationJobStatus>,
-    pub error: Option<String>,
-}
-
-#[cfg(test)]
-mod test {
-    use insta::assert_debug_snapshot;
-    use tempfile::tempdir;
-
-    use super::*;
-
-    async fn register_schema(meta_store: &MetaStore, schema: &'static str) {
-        meta_store
-            .handle(schema.into())
-            .store(DatabaseConfig {
-                is_shared_schema: true,
-                ..Default::default()
-            })
-            .await
-            .unwrap();
-    }
-
-    async fn register_shared(meta_store: &MetaStore, name: &'static str, schema: &'static str) {
-        meta_store
-            .handle(name.into())
-            .store(DatabaseConfig {
-                shared_schema_name: Some(schema.into()),
-                ..Default::default()
-            })
-            .await
-            .unwrap();
-    }
-
-    #[tokio::test]
-    async fn enqueue_migration_job() {
-        let tmp = tempdir().unwrap();
-        let meta_store = MetaStore::new(Default::default(), tmp.path())
-            .await
-            .unwrap();
-        // create 2 shared schema tables
-        register_schema(&meta_store, "schema1").await;
-        register_schema(&meta_store, "schema2").await;
-
-        // create namespaces
-        register_shared(&meta_store, "ns1", "schema1").await;
-        register_shared(&meta_store, "ns2", "schema2").await;
-        register_shared(&meta_store, "ns3", "schema1").await;
-
-        let mut lock = meta_store.inner.lock();
-        // create a migration task
-        lock.register_schema_migration_task(
-            &"schema1".into(),
-            &Program::seq(&["select * from test"]),
-        )
-        .unwrap();
-        let mut stmt = lock.conn.prepare("select * from migration_jobs").unwrap();
-        assert_debug_snapshot!(stmt.query(()).unwrap().next().unwrap().unwrap());
-        stmt.finalize().unwrap();
-
-        let mut stmt = lock
-            .conn
-            .prepare("select * from migration_job_pending_tasks")
-            .unwrap();
-        let mut rows = stmt.query(()).unwrap();
-        assert_debug_snapshot!(rows.next().unwrap().unwrap());
-        assert_debug_snapshot!(rows.next().unwrap().unwrap());
-        assert!(rows.next().unwrap().is_none());
-    }
-
-    #[tokio::test]
-    async fn schema_doesnt_exist() {
-        let tmp = tempdir().unwrap();
-        let meta_store = MetaStore::new(Default::default(), tmp.path())
-            .await
-            .unwrap();
-        // FIXME: the actual error reported here is a shitty constraint error, we should make the
-        // necessary checks beforehand, and return a nice error message.
-        assert!(meta_store
-            .handle("ns1".into())
-            .store(DatabaseConfig {
-                shared_schema_name: Some("schema1".into()),
-                ..Default::default()
-            })
-            .await
-            .is_err());
-    }
-
-    #[tokio::test]
-    async fn pending_job() {
-        let tmp = tempdir().unwrap();
-        let meta_store = MetaStore::new(Default::default(), tmp.path())
-            .await
-            .unwrap();
-
-        register_schema(&meta_store, "schema1").await;
-        register_shared(&meta_store, "ns1", "schema1").await;
-        register_shared(&meta_store, "ns2", "schema1").await;
-        register_shared(&meta_store, "ns3", "schema1").await;
-
-        let job_id = meta_store
-            .register_schema_migration(
-                "schema1".into(),
-                Program::seq(&["create table test (x)"]).into(),
-            )
-            .await
-            .unwrap();
-
-        assert_debug_snapshot!(meta_store.get_next_pending_job().await.unwrap().unwrap());
-
-        let mut tasks = meta_store
-            .get_next_pending_migration_tasks_batch(job_id, MigrationTaskStatus::Enqueued, 10)
-            .await
-            .unwrap();
-        assert_debug_snapshot!(tasks);
-
-        let mut task = tasks.pop().unwrap();
-        task.status = MigrationTaskStatus::Success;
-        meta_store.update_task_status(task, None).await.unwrap();
-
-        assert_debug_snapshot!(meta_store.get_next_pending_job().await.unwrap().unwrap());
-    }
-
-    #[tokio::test]
-    async fn step_job_dry_run_success() {
-        let tmp = tempdir().unwrap();
-        let meta_store = MetaStore::new(Default::default(), tmp.path())
-            .await
-            .unwrap();
-
-        register_schema(&meta_store, "schema1").await;
-        register_shared(&meta_store, "ns1", "schema1").await;
-        register_shared(&meta_store, "ns2", "schema1").await;
-        register_shared(&meta_store, "ns3", "schema1").await;
-        meta_store
-            .register_schema_migration(
-                "schema1".into(),
-                Program::seq(&["create table test (x)"]).into(),
-            )
-            .await
-            .unwrap();
-
-        let job = meta_store.get_next_pending_job().await.unwrap().unwrap();
-        let job = meta_store.job_step_dry_run_success(job).await.unwrap();
-
-        // the job status wasn't updated: there are still tasks that need dry run
-        assert_eq!(job.status, MigrationJobStatus::WaitingDryRun);
-
-        let tasks = meta_store
-            .get_next_pending_migration_tasks_batch(job.job_id, MigrationTaskStatus::Enqueued, 10)
-            .await
-            .unwrap();
-        for mut task in tasks {
-            task.status = MigrationTaskStatus::DryRunSuccess;
-            meta_store.update_task_status(task, None).await.unwrap();
-        }
-
-        let job = meta_store.job_step_dry_run_success(job).await.unwrap();
-        assert_eq!(job.status, MigrationJobStatus::DryRunSuccess);
     }
 }
